refactor(gate): split rule evaluator owner
