+++ docs/architecture-consolidation-plan.md
@@ -82,7 +82,7 @@ fixtures:
 | CLI parser | `crates/tokmd/src/cli/parser.rs` and `crates/tokmd/src/cli/parser/` | `parser.rs` 209; command enum owner 103; global scan args owner 133; shared value-enum owner 235; analyze parser owner 178; context/handoff parser owner 166; cockpit/baseline parser owner 89; diff parser owner 67; gate parser owner 55; sensor parser owner 43; export parser owner 43; badge parser owner 45; module parser owner 36; lang parser owner 26; run parser owner 25; check-ignore parser owner 15; completions parser owner 21; init parser owner 36; tools parser owner 15 | Context, handoff, analyze, cockpit, baseline, diff, gate, sensor, export, badge, module, lang, run, check-ignore, completions, init, tools, command enum, global scan args, and shared value enums now live under parser owner modules; continue command-family splits only when clap snapshots prove behavior is unchanged |
 | CLI config resolution | `crates/tokmd/src/config.rs`, `crates/tokmd/src/config/resolve.rs`, and `crates/tokmd/src/config/resolve/` | `config.rs` 313; resolver barrel 10; lang owner 123; module owner 148; export owner 184; parse owner 35 | Keep config discovery, profile selection, `ConfigContext`, and `ResolvedConfig` in the root config module while lang/module/export CLI-to-receipt argument resolution and string-to-enum parsing live in focused resolver owner modules. Preserve root re-exports. |
 | Gate command | `crates/tokmd/src/commands/gate.rs` and `crates/tokmd/src/commands/gate/` | `gate.rs` 135; policy owner/tests 215; receipt owner/tests 144; render owner 111 | Gate policy loading, ratchet loading, and config-rule conversion live in the policy owner; receipt loading and compute-then-gate preparation live in the receipt owner; text/JSON result rendering lives in the render owner; the command coordinator keeps flow, evaluation, and result combining. Keep `tokmd_gate` proof scoped to all gate command paths. |
-| Gate policy engine | `crates/tokmd-gate/src/evaluate.rs`, `crates/tokmd-gate/src/evaluate/compare.rs`, `crates/tokmd-gate/src/ratchet.rs`, `crates/tokmd-gate/src/ratchet/evaluate.rs`, `crates/tokmd-gate/src/ratchet/policy.rs`, `crates/tokmd-gate/src/numeric.rs`, and `crates/tokmd-gate/src/ratchet/change.rs` | `evaluate.rs` 624; compare owner 84; ratchet coordinator/tests 460; ratchet rule evaluator 175; ratchet policy owner 41; numeric owner 37; ratchet change owner 43 | Policy-rule evaluation and ratchet coordination/tests stay as workflow owners while rule value comparison, single-rule ratchet evaluation, ratchet policy aggregation, shared JSON numeric coercion, and ratchet percentage-change behavior live in private owner modules; strict ratchet evaluation delegates to the configurable evaluator to avoid behavior drift. Keep all gate engine changes under the `tokmd_gate` proof scope. |
+| Gate policy engine | `crates/tokmd-gate/src/evaluate.rs`, `crates/tokmd-gate/src/evaluate/rule.rs`, `crates/tokmd-gate/src/evaluate/compare.rs`, `crates/tokmd-gate/src/ratchet.rs`, `crates/tokmd-gate/src/ratchet/evaluate.rs`, `crates/tokmd-gate/src/ratchet/policy.rs`, `crates/tokmd-gate/src/numeric.rs`, and `crates/tokmd-gate/src/ratchet/change.rs` | `evaluate.rs` 182; policy rule owner/tests 450; compare owner 84; ratchet coordinator/tests 460; ratchet rule evaluator 175; ratchet policy owner 41; numeric owner 37; ratchet change owner 43 | Policy coordination, single-rule evaluation, rule value comparison, and ratchet coordination/tests live in separate workflow/owner modules while single-rule ratchet evaluation, ratchet policy aggregation, shared JSON numeric coercion, and ratchet percentage-change behavior live in private owner modules; strict ratchet evaluation delegates to the configurable evaluator to avoid behavior drift. Keep all gate engine changes under the `tokmd_gate` proof scope. |
 | Sensor command | `crates/tokmd/src/commands/sensor.rs` and `crates/tokmd/src/commands/sensor/` | `sensor.rs` 157; findings owner/tests 344; gate mapping owner/tests 189; output owner/tests 141 | Sensor finding emission, cockpit-evidence-to-envelope gate mapping, and 3-layer output topology now live in owner modules while the command keeps git resolution, cockpit execution, and envelope assembly. |
 | Model aggregation | `crates/tokmd-model/src/lib.rs`, `crates/tokmd-model/src/aggregate.rs`, `crates/tokmd-model/src/children.rs`, `crates/tokmd-model/src/rows.rs`, and `crates/tokmd-model/src/sorting.rs` | `lib.rs` 255; aggregation owner/tests 344; child-language owner/tests 160; row collection owner/tests 411; sorting owner/tests 91 | Report builders, file-row collection, child/embedded language aggregation, in-memory row detection, and row sorting now live in owner modules; future model work should start from concrete product or proof evidence rather than splitting for line count |
 | Scan root handling | `crates/tokmd-scan/src/lib.rs`, `crates/tokmd-scan/src/roots.rs`, and `crates/tokmd-scan/src/path/` | scan facade 534; root validation/rebasing owner 62; bounded path primitives/tests under `path/` | Keep public scan/config/in-memory facades and ignore-pattern expansion in `lib.rs`; keep validated root construction and caller-facing report-path rebasing in `roots.rs`; keep bounded relative/root path primitives under `path/`. Preserve the `model_scan_path_normalization` proof scope. |
